backend/requirements.txt
--extra-index-url https://d33sy5i8bnduwe.cloudfront.net/simple/
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
bcrypt==4.1.3
PyJWT==2.12.1
python-dotenv==1.2.2
pydantic==2.12.5
python-multipart
emergentintegrations
reportlab==4.4.10
sendgrid==6.12.5
